backend/requirements.txt
flask>=3.0
flask-sqlalchemy>=3.1
flask-migrate>=4.0
flask-cors>=5.0
psycopg2-binary>=2.9
PyJWT>=2.8
werkzeug>=3.0
pyscard>=2.0.7
PyYAML>=6.0
reportlab>=4.0
Pillow>=10.0
gunicorn>=22.0
pytest>=8.0
